Adds devise views as haml.
Adds bootstrap to devise views
Makes main nav appear everywhere
Adds some css.

+++ app/assets/javascripts/components/main_nav.js.jsx.coffee
@@ -1,14 +1,25 @@
 @MainNav = React.createClass
+  logoutComponent: ->
+    if !@props.logout_url then return ''
+
+    return `<li>
+      <a href={this.props.logout_url} data-method="delete">Logout</a>
+    </li>`
+
+  usernameDisplay: ->
+    if !@props.username then return
+
+    return `<li>
+      <p className="navbar-text">{this.props.username}</p>
+    </li>`
+
   render: ->
+
     `<nav className="navbar navbar-default">
       <div className="collapse navbar-collapse">
         <ul className="nav navbar-nav navbar-right">
-          <li>
-            <p className="navbar-text">{this.props.username}</p>
-          </li>
-          <li>
-            <a href={this.props.logout_url} data-method="delete">Logout</a>
-          </li>
+          {this.usernameDisplay()}
+          {this.logoutComponent()}
         </ul>
       </div>
     </nav>`
